requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.8.0
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.0.9
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
wrapt==1.14.1
zipp==3.8.0
